+++ src/App.js
@@ -1,13 +1,13 @@
 import './App.css';
 import React from "react";
 import { CustomAppBar } from './components/Header';
-import { Cadastro } from './components/cadastro';
+import { Login } from './components/login';
 
 function App() {
   return (
     <div className="App" style={{height:"100vh", textAlign:"center", backgroundColor:"#202020"}}>
       <CustomAppBar/>
-      <Cadastro/>
+      <Login/>
     </div>
   );
 }
